Split model values into smaller files by category
